+++ refinedstorage2-network/src/main/java/com/refinedmods/refinedstorage2/api/network/impl/security/SecurityNetworkComponentImpl.java
@@ -1,30 +1,35 @@
 package com.refinedmods.refinedstorage2.api.network.impl.security;
 
+import com.refinedmods.refinedstorage2.api.core.CoreValidations;
 import com.refinedmods.refinedstorage2.api.network.node.container.NetworkNodeContainer;
 import com.refinedmods.refinedstorage2.api.network.security.Permission;
 import com.refinedmods.refinedstorage2.api.network.security.SecurityActor;
 import com.refinedmods.refinedstorage2.api.network.security.SecurityDecision;
 import com.refinedmods.refinedstorage2.api.network.security.SecurityDecisionProvider;
 import com.refinedmods.refinedstorage2.api.network.security.SecurityNetworkComponent;
+import com.refinedmods.refinedstorage2.api.network.security.SecurityPolicy;
 
-import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.Set;
 
 public class SecurityNetworkComponentImpl implements SecurityNetworkComponent {
-    private final Set<SecurityDecisionProvider> providers = new HashSet<>();
+    private final Set<SecurityDecisionProvider> providers = new LinkedHashSet<>();
+    private final SecurityPolicy defaultPolicy;
+
+    public SecurityNetworkComponentImpl(final SecurityPolicy defaultPolicy) {
+        this.defaultPolicy = defaultPolicy;
+    }
 
     @Override
     public void onContainerAdded(final NetworkNodeContainer container) {
-        // TODO: use getNode
-        if (container instanceof SecurityDecisionProvider provider) {
+        if (container.getNode() instanceof SecurityDecisionProvider provider) {
             providers.add(provider);
         }
     }
 
     @Override
     public void onContainerRemoved(final NetworkNodeContainer container) {
-        // TODO: use getNode
-        if (container instanceof SecurityDecisionProvider provider) {
+        if (container.getNode() instanceof SecurityDecisionProvider provider) {
             providers.remove(provider);
         }
     }
@@ -32,13 +37,16 @@ public class SecurityNetworkComponentImpl implements SecurityNetworkComponent {
     @Override
     public boolean isAllowed(final Permission permission, final SecurityActor actor) {
         for (final SecurityDecisionProvider provider : providers) {
-            final SecurityDecision decision = provider.isAllowed(permission, actor);
+            final SecurityDecision decision = CoreValidations.validateNotNull(
+                provider.isAllowed(permission, actor),
+                "Security decision provider must not return null"
+            );
             if (decision == SecurityDecision.DENY) {
                 return false;
             } else if (decision == SecurityDecision.ALLOW) {
                 return true;
             }
         }
-        return true;
+        return defaultPolicy.isAllowed(permission);
     }
 }