+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/security/FallbackSecurityCardExtendedMenuProvider.java
@@ -1,8 +1,11 @@
 package com.refinedmods.refinedstorage2.platform.common.security;
 
+import com.refinedmods.refinedstorage2.api.network.security.SecurityPolicy;
+import com.refinedmods.refinedstorage2.platform.api.security.PlatformPermission;
 import com.refinedmods.refinedstorage2.platform.api.support.network.bounditem.SlotReference;
 import com.refinedmods.refinedstorage2.platform.common.content.ContentNames;
 
+import java.util.Set;
 import javax.annotation.Nullable;
 
 import net.minecraft.network.chat.Component;
@@ -13,9 +16,12 @@ import net.minecraft.world.inventory.AbstractContainerMenu;
 class FallbackSecurityCardExtendedMenuProvider extends AbstractSecurityCardExtendedMenuProvider {
     private final SlotReference slotReference;
 
-    FallbackSecurityCardExtendedMenuProvider(final SlotReference slotReference, final SecurityCardModel model) {
-        super(slotReference, model);
-        this.slotReference = slotReference;
+    FallbackSecurityCardExtendedMenuProvider(final SlotReference slotReference,
+                                             final SecurityPolicy securityPolicy,
+                                             final Set<PlatformPermission> dirtyPermissions,
+                                             final SlotReference slotReference1) {
+        super(slotReference, securityPolicy, dirtyPermissions);
+        this.slotReference = slotReference1;
     }
 
     @Override