+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/security/SecurityCardScreen.java
@@ -1,20 +1,12 @@
 package com.refinedmods.refinedstorage2.platform.common.security;
 
-import javax.annotation.Nullable;
-
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.entity.player.Inventory;
-
-import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createTranslation;
 
 public class SecurityCardScreen extends AbstractSecurityCardScreen<SecurityCardContainerMenu> {
     private static final int BOUND_PLAYER_BUTTON_RIGHT_PADDING = 6;
     private static final int BOUND_PLAYER_BUTTON_WIDTH = 80;
-    private static final MutableComponent UNBOUND_TITLE = Component.literal("<")
-        .append(createTranslation("gui", "security_card.unbound"))
-        .append(">");
 
     public SecurityCardScreen(final SecurityCardContainerMenu menu,
                               final Inventory playerInventory,
@@ -25,9 +17,7 @@ public class SecurityCardScreen extends AbstractSecurityCardScreen<SecurityCardC
     @Override
     protected void init(final int rows) {
         super.init(rows);
-        final Component boundToText = menu.getBoundTo() == null
-            ? UNBOUND_TITLE
-            : Component.literal(menu.getBoundTo().name());
+        final Component boundToText = Component.literal(menu.getBoundTo().name());
         final Button boundPlayerButton = Button.builder(boundToText, this::toggleBoundPlayer)
             .pos(leftPos + imageWidth - BOUND_PLAYER_BUTTON_RIGHT_PADDING - BOUND_PLAYER_BUTTON_WIDTH, topPos + 4)
             .size(BOUND_PLAYER_BUTTON_WIDTH, 14)
@@ -39,20 +29,11 @@ public class SecurityCardScreen extends AbstractSecurityCardScreen<SecurityCardC
         if (menu.getPlayers().isEmpty()) {
             return;
         }
-        if (menu.getBoundTo() == null) {
-            setBoundPlayer(button, menu.getPlayers().get(0));
-            return;
-        }
-        final int nextIndex = menu.getPlayers().indexOf(menu.getBoundTo()) + 1;
-        if (nextIndex >= menu.getPlayers().size()) {
-            setBoundPlayer(button, null);
-        } else {
-            setBoundPlayer(button, menu.getPlayers().get(nextIndex));
-        }
-    }
-
-    private void setBoundPlayer(final Button button, @Nullable final SecurityCardContainerMenu.Player player) {
-        menu.changeBoundPlayer(player);
-        button.setMessage(player == null ? UNBOUND_TITLE : Component.literal(player.name()));
+        final SecurityCardContainerMenu.Player currentPlayer = menu.getBoundTo();
+        final int index = menu.getPlayers().indexOf(currentPlayer);
+        final int nextIndex = (index + 1) % menu.getPlayers().size();
+        final SecurityCardContainerMenu.Player nextPlayer = menu.getPlayers().get(nextIndex);
+        menu.changeBoundPlayer(nextPlayer);
+        button.setMessage(Component.literal(nextPlayer.name()));
     }
 }